fix(chipmunk_mring): review fixes — secret wipe, error mapping, docs
H1: wipe l_b and l_x_flat secret material before DAP_DELETE in sign_core
M1: add -EBADMSG/-EIO to s_map_int_err for better diagnostics
M3: fix fold_depth comment to 1+ceil(log2(N)), 2..9
M4: cast coefficient to uint32_t before mask for C11 CT portability
M2: update README_MRNG.md wire-size table with c* qpack (+1408 B)

+++ module/crypto/src/sig/chipmunk/chipmunk_mring_statement.c
@@ -468,7 +468,7 @@ int chipmunk_mring_aggregate_X(chipmunk_poly_t a_X_out[CHIPMUNK_MRING_K_PK],
             const chipmunk_poly_t *l_src =
                 &a_x_flat[i * CHIPMUNK_MRING_K_PK + j];
             for (size_t k = 0u; k < CHIPMUNK_N; ++k) {
-                l_tmp.coeffs[k] = l_src->coeffs[k] & (int32_t)l_mask;
+                l_tmp.coeffs[k] = (int32_t)((uint32_t)l_src->coeffs[k] & l_mask);
             }
             const int rc = chipmunk_poly_add(&a_X_out[j], &a_X_out[j], &l_tmp);
             if (rc != 0) {
